+++ samples/Crabalidator.Samples.Basic/AddressValidator.cs
@@ -1,0 +1,11 @@
+namespace Crabalidator.Samples.Basic
+{
+    public sealed class AddressValidator : CrabValidator<Address>
+    {
+        public AddressValidator()
+        {
+            RuleFor(x => x.PostalCode)
+                .Length(5, 5);
+        }
+    }
+}